+++ src/charts/index.ts
@@ -1,2 +1,4 @@
 export { createExcelWorksheet } from './workbook'
 export { makeXmlCharts } from './xml'
+export { makeXmlChartEx } from './chartex'
+export { chartColorsPartName, chartStylePartName, makeXmlChartColors, makeXmlChartStyle, wantsChartStyleParts } from './style'